Fixed Emerald AI error messages & added "changesForm" boolean
The Emerald AI error was being caused by the actual animation clip itself-- turns out it had an "EmeraldAttackEvent" event tied to it in the animator. I removed this, and I also modified the AI controller script to include a "changesForm" bool, so that it may be used as a more generic AI system for enemies that do not change form.