Puzzle Logic Finished
Puzzle game object can be interacted with. When the puzzle is completed, the enemy in the level will be killed, and another game object will spawn that, upon interaction, will exit the scene to the overworld.
Notes: if you update an object in the scene I would recommend replacing the CanvasLoader event system as well, otherwise you may experience an error for some reason.